[WFCORE-1239] Tighten up the embedded API
Introduce a new factory class EmbeddedProcessFactory to replace the now deprecated EmbeddedServerFactory.
Introduce a new EmbeddedManagedProcess interface as the super-interface for the existing StandaloneServer and HostController interfaces, thus unifying their APIs
Subclass ServerStartException with EmbeddedProcessStartException and then deprecate the superclass.
Move offline CLI to the above.
Misc minor cleanups.
