+++ cli/src/main/java/org/jboss/as/cli/embedded/StopEmbeddedServerHandler.java
@@ -35,10 +35,10 @@ import org.jboss.as.cli.handlers.CommandHandlerWithHelp;
  */
 class StopEmbeddedServerHandler extends CommandHandlerWithHelp {
 
-    private final AtomicReference<EmbeddedServerLaunch> serverReference;
+    private final AtomicReference<EmbeddedProcessLaunch> serverReference;
 
 
-    StopEmbeddedServerHandler(final AtomicReference<EmbeddedServerLaunch> serverReference) {
+    StopEmbeddedServerHandler(final AtomicReference<EmbeddedProcessLaunch> serverReference) {
         super("stop-embedded-server", false);
         assert serverReference != null;
         this.serverReference = serverReference;
@@ -51,14 +51,14 @@ class StopEmbeddedServerHandler extends CommandHandlerWithHelp {
 
     @Override
     protected void doHandle(CommandContext ctx) throws CommandLineException {
-        EmbeddedServerLaunch serverLaunch = serverReference.get();
+        EmbeddedProcessLaunch serverLaunch = serverReference.get();
         if (serverLaunch != null) {
             ctx.disconnectController();
         }
     }
 
-    static void cleanup(final AtomicReference<EmbeddedServerLaunch> serverReference) {
-        EmbeddedServerLaunch serverLaunch = serverReference.get();
+    static void cleanup(final AtomicReference<EmbeddedProcessLaunch> serverReference) {
+        EmbeddedProcessLaunch serverLaunch = serverReference.get();
         if (serverLaunch != null) {
             try {
                 serverLaunch.stop();
